+++ internal/value/nullable.go
@@ -299,7 +299,7 @@ func NullableDyNumberValue(v *string) Value {
 // Nullable makes optional value from nullable type
 // Warning: type interface will be replaced in the future with typed parameters pattern from go1.18
 //
-//nolint:gocyclo
+//nolint:gocyclo, funlen
 func Nullable(t types.Type, v interface{}) Value {
 	switch t {
 	case types.Bool: