+++ internal/xtest/manytimes.go
@@ -38,10 +38,11 @@ func TestManyTimes(t testing.TB, test TestFunc, opts ...TestManyTimesOption) {
 	}
 
 	start := time.Now()
+	var testMutex sync.Mutex
 	for {
 		testCounter++
 		// run test, then check stopAfter for guarantee run test least once
-		runTest(t, test)
+		runTest(t, test, &testMutex)
 
 		if time.Since(start) > options.stopAfter || t.Failed() {
 			return
@@ -60,11 +61,12 @@ func TestManyTimesWithName(t *testing.T, name string, test TestFunc) {
 
 type TestFunc func(t testing.TB)
 
-func runTest(t testing.TB, test TestFunc) {
+func runTest(t testing.TB, test TestFunc, testMutex *sync.Mutex) {
 	t.Helper()
 
 	tw := &testWrapper{
 		TB: t,
+		m:  testMutex,
 	}
 
 	defer tw.doCleanup()
@@ -75,7 +77,7 @@ func runTest(t testing.TB, test TestFunc) {
 type testWrapper struct {
 	testing.TB
 
-	m       sync.Mutex
+	m       *sync.Mutex
 	logs    []logRecord
 	cleanup []func()
 }